move s3 utils to a separate storage backend folder
