+++ src/routes/object/deleteObjects.ts
@@ -5,7 +5,7 @@ import { AuthenticatedRequest, Obj } from '../../types/types'
 import { getPostgrestClient, transformPostgrestError } from '../../utils'
 import { getConfig } from '../../utils/config'
 import { createDefaultSchema } from '../../utils/generic-routes'
-import { deleteObjects, initClient } from '../../utils/s3'
+import { deleteObjects, initClient } from '../../backend/s3'
 
 const { region, projectRef, globalS3Bucket, globalS3Endpoint } = getConfig()
 const client = initClient(region, globalS3Endpoint)
